requirements.txt
# packages in environment at /Users/wli17/opt/anaconda3/envs/dMEGA:
#
# Name                    Version                   Build  Channel
blas                      1.0                         mkl  
bottleneck                1.3.5           py310h4e76f89_0  
bzip2                     1.0.8                h1de35cc_0  
ca-certificates           2022.4.26            hecd8cb5_0  
certifi                   2022.6.15       py310hecd8cb5_0  
cffi                      1.15.0          py310hc55c11b_1  
future                    0.18.2          py310hecd8cb5_1  
intel-openmp              2021.4.0          hecd8cb5_3538  
joblib                    1.1.0              pyhd3eb1b0_0  
libcxx                    12.0.0               h2f01273_0  
libffi                    3.3                  hb1e8313_2  
libgfortran               3.0.1                h93005f0_2  
llvm-openmp               12.0.0               h0dcd299_1  
mkl                       2021.4.0           hecd8cb5_637  
mkl-service               2.4.0           py310hca72f7f_0  
mkl_fft                   1.3.1           py310hf879493_0  
mkl_random                1.2.2           py310hc081a56_0  
ncurses                   6.3                  hca72f7f_3  
ninja                     1.10.2               hecd8cb5_5  
ninja-base                1.10.2               haf03e11_5  
numexpr                   2.8.3           py310hdcd3fac_0  
numpy                     1.22.3          py310hdcd3fac_0  
numpy-base                1.22.3          py310hfd2de13_0  
openssl                   1.1.1q               hca72f7f_0  
packaging                 21.3               pyhd3eb1b0_0  
pandas                    1.4.3           py310he9d5cce_0  
pip                       22.1.2          py310hecd8cb5_0  
pycparser                 2.21               pyhd3eb1b0_0  
pyparsing                 3.0.4              pyhd3eb1b0_0  
python                    3.10.4               hdfd78df_0  
python-dateutil           2.8.2              pyhd3eb1b0_0  
pytorch                   1.10.2          cpu_py310h30e64cd_0  
pytz                      2022.1          py310hecd8cb5_0  
readline                  8.1.2                hca72f7f_1  
scikit-learn              1.1.1           py310he9d5cce_0  
scipy                     1.7.3           py310h3dd3380_0  
setuptools                61.2.0          py310hecd8cb5_0  
six                       1.16.0             pyhd3eb1b0_1  
sqlite                    3.38.5               h707629a_0  
threadpoolctl             2.2.0              pyh0d69192_0  
tk                        8.6.12               h5d9f67b_0  
typing-extensions         4.1.1                hd3eb1b0_0  
typing_extensions         4.1.1              pyh06a4308_0  
tzdata                    2022a                hda174b7_0  
wheel                     0.37.1             pyhd3eb1b0_0  
xz                        5.2.5                hca72f7f_1  
zlib                      1.2.12               h4dc903c_2  
